kick file and map parsing into its own package
